==> 03-Notebooks/04-CPI.ipynb <==
import pandas as pd
from fredapi import Fred

## Obtención de datos de FRED

fred = Fred(api_key='a7eca89fdf2905baea21d67b942c9ef7')
cpi=fred.get_series('CPIAUCNS')

cpi = pd.DataFrame(data=cpi, columns=['Value'])

## Creación del pickle para no llamar a la API siempre

#cpi.to_pickle(r'..\02-Data\04-CPI\01-CPI_USD_1913-01-01_2021-06-01_monthly_fred.pkl')
cpi = pd.read_pickle(r'..\02-Data\04-CPI\01-CPI_USD_1913-01-01_2021-06-01_monthly_fred.pkl')
cpi

cpi.to_csv(r'..\02-Data\04-CPI\01-CPI_USD_1913-01-01_2021-06-01_monthly_fred.csv')
