==> 03-Notebooks/05-Inflation.ipynb <==
import pandas as pd

### Se lee el pickle de CPI para calcular los cambios anuales y mensuales para inflacion

cpi = pd.read_pickle(r'..\02-Data\04-CPI\01-CPI_USD_1913-01-01_2021-06-01_monthly_fred.pkl')

inflation_monthly = cpi.pct_change(12).dropna()

inflation_monthly

inflation_monthly.to_pickle(r'..\02-Data\05-Inflation\01_Inflation_percentage_1913-01-01_2021-06-01_monthly_fred.pkl')

inflation_monthly.to_parquet(r'..\02-Data\05-Inflation\01_Inflation_percentage_1913-01-01_2021-06-01_monthly_fred.')

inflation_monthly.to_parquet(r'..\02-Data\05-Inflation\01_Inflation_percentage_1913-01-01_2021-06-01_monthly_fred.parquet')

inflation_anually = cpi.groupby(cpi.index.strftime('%Y')).tail(1).pct_change(1).dropna()
inflation_anually

inflation_anually.to_pickle(r'..\02-Data\05-Inflation\02_Inflation_percentage_1913-01-01_2021-06-01_anually_fred.pkl')
