==> 03-Notebooks/06-M2.ipynb <==
from fredapi import Fred
import pandas as pd
from datetime import *
from pandas.tseries.offsets import MonthEnd

m2_prev = pd.read_excel(r'..\01-Raw_data\03-M2\m2_billions.usd_1867_1970_anually.xlsx')

m2_prev = m2_prev[['Year', 'M2']]
m2_prev = m2_prev.dropna()
m2_prev.rename(columns={"M2": "Value"}, inplace=True)
m2_prev['Year'] = m2_prev['Year'].astype(int)
m2_prev['Year'] = pd.to_datetime(m2_prev['Year'], format='%Y')
m2_prev = m2_prev.set_index('Year')

mask = (m2_prev.index < datetime.strptime('1959-01-01', '%Y-%m-%d'))
m2_prev = m2_prev.loc[mask]
m2_prev.index = m2_prev.index.map(lambda x: pd.datetime(x.year, 12, x.day))
m2_prev.index = pd.to_datetime(m2_prev.index, format='%Y') + MonthEnd(1)

m2_prev

fred = Fred(api_key='a7eca89fdf2905baea21d67b942c9ef7')
m2=fred.get_series('M2SL')
m2

m2 = pd.DataFrame(data=m2, columns=['Value'])
m2 = m2.asfreq(freq='YS', method='ffill')
m2

m2_def = m2.append(m2_prev).sort_index().reset_index().set_index('index')
m2_def

m2_def.to_pickle(r'..\02-Data\06-M2\02_M2_USD_1867-12-31-2021-01-01_yearly_fred.pkl')

m2_def = pd.read_pickle(r'..\02-Data\06-M2\02_M2_USD_1867-12-31-2021-01-01_yearly_fred.pkl')

m2_pct = m2_def.pct_change(1).dropna()

m2_pct.to_pickle(r'..\02-Data\06-M2\01_M2_percentage_1868-12-31-2021-01-01_yearly_fred.pkl')

m2_pct.to_parquet(r'..\02-Data\06-M2\01_M2_percentage_1868-12-31-2021-01-01_yearly_fred.parquet')

m2_pct

m2_prev_pct = m2_prev.pct_change(1)
