==> 03-Notebooks/07-Nominal_interest_rate.ipynb <==
from fredapi import Fred
import pandas as pd

fred = Fred(api_key='a7eca89fdf2905baea21d67b942c9ef7')
interest_rates=fred.get_series('DFF')
interest_rates

interest_rates = pd.DataFrame(data=interest_rates, columns=['Value'])

interest_rates

interest_rates.to_pickle(r'..\02-Data\07-Interest_rates\01-InterestRates_percentage_1954-07-01_2021-07-22_daily_fred.pkl')

interest_rates_mensual = interest_rates.asfreq('M')

interest_rates_mensual.to_pickle(r'..\02-Data\07-Interest_rates\02-InterestRates_percentage_1954-07-01_2021-07-22_monthly_fred.pkl')

interest_rates_anual = interest_rates.asfreq('Y')
interest_rates_anual = interest_rates_anual.append(interest_rates_mensual.tail(1))

interest_rates_anual

interest_rates_anual.to_pickle(r'..\02-Data\07-Interest_rates\03-InterestRates_percentage_1954-07-01_2021-07-22_anually_fred.pkl')

## Tasa de interés a 10 años

fred = Fred(api_key='a7eca89fdf2905baea21d67b942c9ef7')
interest_rates_10_year=fred.get_series('DGS10')
interest_rates_10_year

interest_rates_10_year = pd.DataFrame(data=interest_rates_10_year, columns=['Value'])

interest_rates_10_year = interest_rates_10_year.dropna()

interest_rates_10_year.to_pickle(r'..\02-Data\07-Interest_rates\04-10YearInterestRates_percentage_1962-01-01_2021-07-22_daily_fred.pkl')

## 10 year treasury inflation-indexed security

fred = Fred(api_key='a7eca89fdf2905baea21d67b942c9ef7')
inflation_indexed_security=fred.get_series('DFII10')
inflation_indexed_security

inflation_indexed_security = pd.DataFrame(data=inflation_indexed_security, columns=['Value'])

inflation_indexed_security = inflation_indexed_security.dropna()/100

inflation_indexed_security.to_pickle(r'..\02-Data\07-Interest_rates\05-10YearTreasuryInflationIndexedSecurity_percentage_2003-01-02_2021-07-22_daily_fred.pkl')

inflation_indexed_security.to_parquet(r'..\02-Data\07-Interest_rates\05-10YearTreasuryInflationIndexedSecurity_percentage_2003-01-02_2021-07-22_daily_fred.parquet')
