==> 03-Notebooks/08-Real_interest_rates.ipynb <==
import pandas as pd
from datetime import *

### Tasa de interés real = Tasa de interés nominal - Inflación

### Datos mensuales

inflation_monthly = pd.read_pickle(r'..\02-Data\05-Inflation\01_Inflation_percentage_1913-01-01_2021-06-01_monthly_fred.pkl')
interest_rate_monthly = pd.read_pickle(r'..\02-Data\07-Interest_rates\02-InterestRates_percentage_1954-07-01_2021-07-22_monthly_fred.pkl')

inflation_monthly

interest_rate_monthly

mask = (inflation_monthly.index >= datetime.strptime('1954-07-01', '%Y-%m-%d'))
inflation_monthly= inflation_monthly.loc[mask]
inflation_monthly

real_interest_rate_monthly = interest_rate_monthly.set_index(inflation_monthly.index).subtract(inflation_monthly, fill_value=0)

real_interest_rate_monthly

real_interest_rate_monthly.to_pickle(r'..\02-Data\08-Real_interest_rates\01-RealInterestRates_percentage_1954-07-01_2021-06-01_monthly_fred.pkl')

real_interest_rate_monthly.to_parquet(r'..\02-Data\08-Real_interest_rates\01-RealInterestRates_percentage_1954-07-01_2021-06-01_monthly_fred.parquet')

real_interest_rate_monthly

### Datos anuales

inflation_yearly = pd.read_pickle(r'..\02-Data\05-Inflation\02_Inflation_percentage_1913-01-01_2021-06-01_anually_fred.pkl')
interest_rate_yearly = pd.read_pickle(r'..\02-Data\07-Interest_rates\03-InterestRates_percentage_1954-07-01_2021-07-22_anually_fred.pkl')

inflation_yearly

mask_2 = (inflation_yearly.index >= datetime.strptime('1954-12-01', '%Y-%m-%d'))
inflation_yearly= inflation_yearly.loc[mask_2]
inflation_yearly

interest_rate_yearly

real_interest_rate_yearly = interest_rate_yearly.set_index(inflation_yearly.index).subtract(inflation_yearly, fill_value=0)

real_interest_rate_yearly.to_pickle(r'..\02-Data\08-Real_interest_rates\02-RealInterestRates_percentage_1954-12-01_2021-06-01_yearly_fred.pkl')

real_interest_rate_yearly
