==> 03-Notebooks/09-Inflation_Europe.ipynb <==
import requests
import pandas as pd
import io
from pandas.tseries.offsets import MonthEnd

# Building blocks for the URL
entrypoint = 'https://sdw-wsrest.ecb.europa.eu/service/' # Using protocol 'https'
resource = 'data'           # The resource for data queries is always'data'
flowRef ='ICP'              # Dataflow describing the data that needs to be returned, exchange rates in this case
key = 'M.U2.N.000000.4.ANR'    # Defining the dimension values, explained below

# Define the parameters
'''parameters = {
    'startPeriod': '2000-01-01',  # Start date of the time series
    'endPeriod': '2018-10-01'     # End of the time series
}'''

request_url = entrypoint + resource + '/'+ flowRef + '/' + key

response = requests.get(request_url, headers={'Accept': 'text/csv'})

inflation_europe = pd.read_csv(io.StringIO(response.text))

inflation_europe = inflation_europe[['TIME_PERIOD', 'OBS_VALUE']]

inflation_europe.rename(columns={
    'TIME_PERIOD': 'Date',
    'OBS_VALUE': 'Value'
}, inplace=True)

inflation_europe.set_index('Date', inplace=True)

inflation_europe.index = pd.to_datetime(inflation_europe.index, format='%Y-%m')

inflation_europe.index = pd.to_datetime(inflation_europe.index, format='%Y') + MonthEnd(1)

inflation_europe

inflation_europe.to_pickle(r'..\02-Data\09-Inflation_Europe\01-Inflation_percentage_1997-01-31_2021-06-30_monthly_ecb.pkl')
