==> 03-Notebooks/11-Real_interest_rates_Europe.ipynb <==
import pandas as pd
from io import StringIO

inflation_europe = pd.read_pickle(r'..\02-Data\09-Inflation_Europe\01-Inflation_percentage_1997-01-31_2021-06-30_monthly_ecb.pkl')
interest_rates_europe = pd.read_pickle(r'..\02-Data\10-Interest_rates_Europe\01-Interest_rates_corporations_2003-01-31_2021-05-31_monthly_ecb.pkl')

inflation_europe.drop(inflation_europe.tail(1).index, inplace=True)

inflation_europe

interest_rates_europe

real_interest_rates= interest_rates_europe.subtract(inflation_europe).dropna()

real_interest_rates

real_interest_rates.to_pickle(r'..\02-Data\11-Real_interest_rates_Europe\01-Real_interest_rates_2003-01-31_2021-05-31_monthly_ecb.pkl')

real_interest_rates.to_parquet(r'..\02-Data\11-Real_interest_rates_Europe\01-Real_interest_rates_2003-01-31_2021-05-31_monthly_ecb.parquet')
